backend/events/views.py
from rest_framework import viewsets
from .models import EventCategory, Event
from .serializers import EventCategorySerializer, EventSerializer

class EventCategoryViewSet(viewsets.ReadOnlyModelViewSet):
    queryset = EventCategory.objects.all()
    serializer_class = EventCategorySerializer

class EventViewSet(viewsets.ReadOnlyModelViewSet):
    queryset = Event.objects.all()
    serializer_class = EventSerializer
